Improve the code structure (#10)
