+++ pysimenv/multicopter/__init__.py
@@ -1,7 +1,8 @@
 """
 Description for Package
 """
-from pysimenv.multicopter.model import EffectorModel, Mixer, MulticopterDyn, QuadXEffector, QuadXMixer, ActuatorFault
+from pysimenv.multicopter.base import MulticopterDyn, EffectorModel, Mixer
+from pysimenv.multicopter.model import QuadBase, QuadXEffector, QuadXMixer, ActuatorFault
 from pysimenv.multicopter.control import BSControl, SMAttControl, FLVelControl, QuaternionAttControl, QuaternionPosControl
 from pysimenv.multicopter.estimator import FixedTimeFaultEstimator
