+++ pysimenv/multicopter/test_sm_att_control.py
@@ -1,7 +1,8 @@
 import numpy as np
 from pysimenv.core.simulator import Simulator
 from pysimenv.common.orientation import euler_angles_to_rotation
-from pysimenv.multicopter.model import MulticopterDyn, QuadBase
+from pysimenv.multicopter.base import MulticopterDyn
+from pysimenv.multicopter.model import QuadBase
 from pysimenv.multicopter.control import SMAttControl
